Using the Array Find Method
Created a variable called superBowlWin and assigned it to an arrow function that takes in one parameter(array) within the arrow function we careate a variable called year and assign it to the rcords and iterate through it to try and find if the given team has ever won a superbowl. WWe then use the ternary operator and if through the iteration there is a value that returns true that the team has won it returns that year. If the team has never won then it will return undefined